pyproject.toml
[project]
name = "my-fastapi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlmodel",
    "psycopg[binary]",
    "sqlalchemy[asyncio]",
    "fastapi[all]==0.115.13",
    "asyncpg>=0.30.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "pydantic-settings==2.9.1",
    "alembic>=1.16.2",
    "psycopg2-binary>=2.9.10",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.6.15",
    "click==8.2.1",
    "colorama==0.4.6",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi-cli==0.0.7",
    "grpcio==1.70.0",
    "grpcio-tools==1.70.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "orjson==3.10.18",
    "protobuf==5.29.3",
    "pydantic==2.11.7",
    "pydantic-extra-types==2.10.5",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rich==14.0.0",
    "rich-toolkit==0.14.7",
    "setuptools==75.8.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "typer==0.16.0",
    "typing-inspection==0.4.1",
    "typing-extensions==4.14.0",
    "ujson==5.10.0",
    "uv==0.7.6",
    "uvicorn==0.34.3",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]
